Fixed signup_login functional tests, added combat func tests
