+++ .gitignore
@@ -12,6 +12,8 @@ db.sqlite3-journal
 media
 local_test_settings.py
 dev_settings.py
+chromedriver.exe
+geckodriver.exe
 
 # If your build process includes running collectstatic, then you probably don't need or want to include staticfiles/
 # in your Git repository. Update and uncomment the following line accordingly.
