+++ templates/game/game.html
@@ -41,11 +41,11 @@
                 <p class="text-success"> You hit the {{ current_enemy }} for {{ dmg_delt_player }} damage and killed it!</p>
                 <p class="text-success">You looted {{ looted_power_crystals }} power crystals and {{ looted_gold }} gold peices!
             {% endif %}
-            <form method="POST", class="form-group">
+            <form method="POST", class="form-group" id="fight_form">
                 {% csrf_token %}
                 <br>
                 <label for="enemy_dropdown">Choose an enemy to fight:</label>
-                    <select name="enemy_dropdown" id="enemy_dropdown">
+                    <select name="enemy_dropdown" id="enemy_dropdown_list">
                         {% if dmg_delt_enemy %}
                             {% for enemy in enemies %}
                                 <option value="{{ enemy }}">{{ enemy.name }}</option>
